Start transition to Crinja and Bulma templating. Based on new mockups for 2019. Bulma is a better baseline to style from and provides a softer feel than Bootstrap, while also providing easier customization and extensibility.
Jinja is a runtime templating system (as opposed to Slang and ECR's compile time templating), meaning the templates are evaluated every time a page is loaded, rather than being compiled into the application binary itself. While this limits portability to a small extent, it also drastically reduces compile times, which have been out-of-hand and distracting for a while now. The conversion process will take a while, so both rendering systems are currently in place to avoid breaking any existing pages.