Switch to Orion for routing instead of Kemal. Orion provides far more powerful routing control, superior middleware instrumentation, and a class-based controller interface that feels less magical and makes a lot of things cleaner to work with around the application (i.e., checking admin permissions and redirecting at the router level).
This commit also clears out all of the remaining `slang` layouts and controllers that were using them, as they are not currently necessary and bloated compile times. The compiles still aren't great, but they're about as good as possible given the size of this application now with all dependencies included.