nested routing with new admin views
Turns out preact router _does_ do nested routing, it just doesn't explain it well.

The new admin lives in a top-level `admin` folder under app, and essentially acts as its own application with its own modules and router and everything, but still uses the same store as the rest of the app, with everything under the `admin` key of the main reducer. This is abstracted away by the selectors so it can change easily in the future. This also means it has access to the global state if needed (though it shouldn't use this much).