Get `mcsn_supervisor` working to manage running processes.
`mcsn_supervisor` is just an installation of `supervisord` with a configuration set up for managing mcsn's services.

It's not a build process or dependency manager yet, but at least handles the more necessary part of running the applications. And it handles live reloading in a development environment, so everything effectively becomes hands-off.