Set up mcsn CLI to prepare for build steps
The CLI structure finally makes some sense. Services will be able to define themselves as a whole within a single file and be registered under the main CLI command.

I'm not sure how to deal with directory navigation and all that yet. As it stands, the virtualenv for the project has to be activated (bleh) and the user _must_ be in this project's directory for any script to work properly. Ideally it would just work as a symlink-able executable from anywhere in the system, kind of like `systemd`.