add api to mcsn_cli and supervisorctl
The CLI command was relatively simple. It will take care of installing tooling and dependencies with `mcsn api deps`, and even builds a local `sentry` executable to manage live reloading. `mcsn api build` will build a command line version of the CLI for `mcsn_api`, and `mcsn api start` will run the sentry instance in the terminal.

Getting supervisor to work with sentry was a bear. Because `sentry` is an out-of-process reloader, it ends up spawning child processes, which supervisor by default doesn't know how to handle. After a lot of searching, there are `stopasgroup` and `killasgroup` configurations which tell supervisor to send signals to the whole process hierarchy, not just the PID it spawned. For good measure, I also modified the local `sentry` source to use `Process.fork` instead of `Process.run`, which should ensure that child relationship always exists.